Add Density

+++ LithuaniaInNumbers/src/main/java/org/LithuaniaInNumbers/Models/People/AverageAge.java
@@ -2,12 +2,12 @@ package org.LithuaniaInNumbers.Models.People;
 
 public class AverageAge {
 //	public int age;
-	public float year;
+	public int year;
 	public float total;
 	public float men;
 	public float women;
 	
-	public AverageAge(float total, float men, float women, float year) {
+	public AverageAge(float total, float men, float women, int year) {
 //		this.age = age;
 		this.year = year;
 		this.total = total;
